added example classes
added
